Further refactor shop functions into their own file
