+++ source/shops.hpp
@@ -1,0 +1,12 @@
+#pragma once
+#include "item.hpp"
+
+#include <vector>
+
+extern void SetVanillaShopItems();
+extern int GetRandomShopPrice();
+extern void PlaceShopItems();
+extern int GetShopsanityReplaceAmount();
+extern void ShuffleShop(std::vector<Item>& ShopItems, std::vector<int> indicesToExclude);
+
+extern std::vector<Item> ShopItems;
